Should have scaled the game to work better with box2D. Box2D likes small units/measurements for collision and movement.

+++ core/src/com/mygdx/game/behaviourtree/action/FindPath.java
@@ -149,7 +149,7 @@ public class FindPath extends LeafTask {
             node.H = 0;
         }
 
-        int squareSize = this.blackBoard.colonyGrid.getSquareSize();
+        float squareSize = this.blackBoard.colonyGrid.getSquareSize();
         //If a path was found, record the path.
         if(found) {
             currNode = target;
